Release 23.1.0
Release notes are available at: https://docs.dremio.com/release-notes/

+++ plugins/elasticsearch/src/main/java/com/dremio/plugins/elastic/execution/WriteHolders.java
@@ -222,7 +222,7 @@ class WriteHolders {
       if (token == JsonToken.VALUE_NUMBER_FLOAT || token == JsonToken.VALUE_NUMBER_INT) {
         writer.writeFloat8(parser.getDoubleValue());
       } else {
-        writer.writeFloat8(Float.parseFloat(parser.getValueAsString()));
+        writer.writeFloat8(Double.parseDouble(parser.getValueAsString()));
       }
     }
   }
